pyproject.toml
[project]
name = "tg-bidlo-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "EgorLozh",email = "lozhkinyegor@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama (>=0.5.4,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
